requirements.txt
argparse
python-dateutil
Faker
pytest